simplesub: split simplesub.go into focused files
Pure code-organization refactor (no behavior change). Replace the single
~700-line simplesub.go with focused files in the same package:

- doc.go        package overview + source-layout guide
- polarity.go   the Polarity enum
- types.go      the SimpleType representation + levelOf
- constrain.go  constrain(lhs <: rhs) + extrusion + describe helpers
- scheme.go     let-polymorphism: schemes, instantiate, freshenAbove
- infer.go      the expression IR, typeTerm, and public Infer/Render
- simplify.go   occurrence analysis + co-occurrence variable merging
- coalesce.go   coalescing a SimpleType into a type_system.Type
